Move custom validation annotation for PrimeNumberRequest.
