+++ src/main/java/com/example/prime_number_service/model/PrimeNumberRequest.java
@@ -1,7 +1,7 @@
 package com.example.prime_number_service.model;
 
 import com.example.prime_number_service.constant.PrimeNumberAlgorithm;
-import com.example.prime_number_service.validation.ValidPrimeNumberRequest;
+import com.example.prime_number_service.annotation.ValidPrimeNumberRequest;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
